# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sChat REST server",
        "title": "sChat Server API",
        "contact": {
            "name": "mykola.sainyuk@gmail.com",
            "url": "https://www.linkedin.com/in/mykola-sainiuk-3b03168b/"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/api/chat/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated list of User chats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "List chats paginated",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "limit",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ChatOutputDto"
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/list/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Unpaginated list of all user chats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "List all chats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ChatOutputDto"
                            }
                        }
                    }
                }
            }
        },
        "/api/chat/new": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Establish new chat for two users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Create new chat",
                "parameters": [
                    {
                        "description": "New chat input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddChatInputDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not found user",
                        "schema": {
                            "$ref": "#/definitions/httpexp.HttpExp"
                        }
                    }
                }
            }
        },
        "/api/chat/{chatId}/clear": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete all messages from chat",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Clear chat",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/message/{chatId}/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated list of Chat messages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "List messages paginated",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "limit",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.MessageOutputDto"
                            }
                        }
                    }
                }
            }
        },
        "/api/message/{chatId}/list/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Unpaginated list of all chat messages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "List all messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.MessageOutputDto"
                            }
                        }
                    }
                }
            }
        },
        "/api/message/{chatId}/new": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add new message into the chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "Write new message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chatId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New contact input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewMessageInputDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.NewMessageOutputDto"
                        }
                    },
                    "404": {
                        "description": "Not found user",
                        "schema": {
                            "$ref": "#/definitions/httpexp.HttpExp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddChatInputDto": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "chatName": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
...
        "dto.ChatOutputDto": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "iconUri": {
                    "type": "string"
                },
                "lastMessage": {
                    "type": "string"
                },
                "muted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.MessageOutputDto": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "chat": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "received": {
                    "type": "boolean"
                },
                "sent": {
                    "type": "boolean"
                },
                "system": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "dto.NewMessageInputDto": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "image": {
                    "type": "string"
                },
                "text": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "dto.NewMessageOutputDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "httpexp.HttpExp": {
            "description": "common 422 response",
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}